Update README.md
This was written from a backend perspective, but most descriptions attempt to describe the project in full-stack terms. Some areas will need to be changed by FE, specifically the Frontend Repository Installation section. The gems utilized descriptions can be updated to whatever is appropriate for FE or removed. I have the template in for the FE tech stack so when someone has freetime, send me your stack and I will make an identical diagram for this README. Links should all be updated, but I recommend double checking. 

If there is anything y'all would like to add to the summary sections (About The Project), please let me know! The intent was for these to be identical between both READMEs.